+++ project2-branch2/src/main/java/com/example/bookstore/services/implementations/PublisherServiceImpl.java
@@ -3,6 +3,8 @@ package com.example.bookstore.services.implementations;
 import com.example.bookstore.dto.Publisher.PublisherRequestDto;
 import com.example.bookstore.dto.Publisher.PublisherResponseDto;
 import com.example.bookstore.dto.Publisher.PublisherUpdateDto;
+import com.example.bookstore.entities.Book;
+import com.example.bookstore.exceptions.ResourceNotFoundException;
 import com.example.bookstore.repository.BookRepo;
 import com.example.bookstore.repository.PublisherRepo;
 
@@ -14,6 +16,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+
 @Service
 public class PublisherServiceImpl implements PublisherService {
     private final PublisherRepo publisherRepo;
@@ -32,41 +35,52 @@ public class PublisherServiceImpl implements PublisherService {
 
     @Override
     public Optional<PublisherRequestDto> getByID(Long id) {
-        return publisherRepo.findById(id).map(Publisher::convertPublisherToRequestDto);
+        if (publisherRepo.existsById(id))
+            return publisherRepo.findById(id).map(Publisher::convertPublisherToRequestDto);
+        else
+            throw new ResourceNotFoundException(String.format("The publisher with ID: %d is not found or doesn't exist", id));
     }
 
     @Override
     public void deleteByID(Long id) {
-        publisherRepo.deleteById(id);
+        if (publisherRepo.existsById(id))
+            publisherRepo.deleteById(id);
+        else throw
+                new ResourceNotFoundException(String.format("The publisher with ID: %d is not found or already deleted", id));
     }
 
     @Override
     public PublisherResponseDto create(PublisherResponseDto publisherResponseDto) {
-        Publisher publisher= publisherResponseDto.convertPublisherRequestDtoToEntity();
-        Publisher publisherCreated=publisherRepo.save(publisher);
+        Publisher publisher = publisherResponseDto.convertPublisherRequestDtoToEntity();
+        Publisher publisherCreated = publisherRepo.save(publisher);
         return publisherCreated.convertPublisherToResponseDto();
     }
+
     @Override
     public void update(PublisherUpdateDto publisherUpdateDto) {
         Publisher existingPublisher;
-        val books =bookRepo.findAllByIdIn(publisherUpdateDto.getPublishedBooksIds());
+        List<Book> books = bookRepo.findAllByIdIn(publisherUpdateDto.getPublishedBooksIds());
 
-        Publisher publisher= publisherUpdateDto.convertPublisherUpdateDtoToEntity(books);
+        Publisher publisher = publisherUpdateDto.convertPublisherUpdateDtoToEntity(books);
 
-            existingPublisher = publisherRepo.findById(publisher.getId()).orElseThrow();
-            existingPublisher.setName(publisher.getName());
-            existingPublisher.setId(publisher.getId());
-            existingPublisher.setPublishedBooksList(publisher.getPublishedBooksList());
-            publisherRepo.save(publisher);
+        existingPublisher = publisherRepo.findById(publisher.getId()).orElseThrow(() ->
+                new ResourceNotFoundException(String.format("The publisher with ID: %d is not found or doesn't exist", publisherUpdateDto.getId())));
+        existingPublisher.setName(publisher.getName());
+        existingPublisher.setId(publisher.getId());
+        existingPublisher.setPublishedBooksList(publisher.getPublishedBooksList());
+        publisherRepo.save(publisher);
 
 
     }
 
     @Override
     public List<PublisherRequestDto> getByNameContaining(String name) {
-        return publisherRepo
-                .findByNameIsContainingIgnoreCase(name)
-                .stream().map(Publisher::convertPublisherToRequestDto)
-                .toList();
+        if (publisherRepo.existsByNameIsContainingIgnoreCase(name))
+            return publisherRepo
+                    .findByNameIsContainingIgnoreCase(name)
+                    .stream().map(Publisher::convertPublisherToRequestDto)
+                    .toList();
+        else throw
+                new ResourceNotFoundException(String.format("The book with name %s is not found or doesn't exist", name));
     }
 }
